+++ src/SeerAssemblyPreferenceDialog.cpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2021 Ernie Pasveer <epasveer@att.net>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 #include "SeerAssemblyPreferenceDialog.h"
 #include <QtCore/QSettings>
 #include <QtCore/QDebug>