+++ src/SeerStackLocalsBrowserWidget.cpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2021 Ernie Pasveer <epasveer@att.net>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 #include "SeerStackLocalsBrowserWidget.h"
 #include "SeerUtl.h"
 #include <QtWidgets/QTreeWidget>
@@ -151,10 +155,12 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
     QAction* addVariableLoggerAsteriskExpressionAction;
     QAction* addVariableLoggerAmpersandExpressionAction;
     QAction* addVariableLoggerAsteriskAmpersandExpressionAction;
+    QAction* addVariableLoggerObjcExpressionAction;
     QAction* addVariableTrackerExpressionAction;
     QAction* addVariableTrackerAsteriskExpressionAction;
     QAction* addVariableTrackerAmpersandExpressionAction;
     QAction* addVariableTrackerAsteriskAmpersandExpressionAction;
+    QAction* addVariableTrackerObjcExpressionAction;
     QAction* addMemoryVisualizerAction;
     QAction* addMemoryAsteriskVisualizerAction;
     QAction* addMemoryAmpersandVisualizerAction;
@@ -168,26 +174,28 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
     QAction* addStructAsteriskVisualizerAction;
     QAction* addStructAmpersandVisualizerAction;
 
-    addVariableLoggerExpressionAction                   = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addVariableLoggerAsteriskExpressionAction           = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addVariableLoggerAmpersandExpressionAction          = new QAction(QString("\"&&%1\"").arg(item->text(0)));
-    addVariableLoggerAsteriskAmpersandExpressionAction  = new QAction(QString("\"*&&%1\"").arg(item->text(0)));
-    addVariableTrackerExpressionAction                  = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addVariableTrackerAsteriskExpressionAction          = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addVariableTrackerAmpersandExpressionAction         = new QAction(QString("\"&&%1\"").arg(item->text(0)));
-    addVariableTrackerAsteriskAmpersandExpressionAction = new QAction(QString("\"*&&%1\"").arg(item->text(0)));
-    addMemoryVisualizerAction                           = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addMemoryAsteriskVisualizerAction                   = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addMemoryAmpersandVisualizerAction                  = new QAction(QString("\"&&%1\"").arg(item->text(0)));
-    addArrayVisualizerAction                            = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addArrayAsteriskVisualizerAction                    = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addArrayAmpersandVisualizerAction                   = new QAction(QString("\"&&%1\"").arg(item->text(0)));
-    addMatrixVisualizerAction                           = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addMatrixAsteriskVisualizerAction                   = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addMatrixAmpersandVisualizerAction                  = new QAction(QString("\"&&%1\"").arg(item->text(0)));
-    addStructVisualizerAction                           = new QAction(QString("\"%1\"").arg(item->text(0)));
-    addStructAsteriskVisualizerAction                   = new QAction(QString("\"*%1\"").arg(item->text(0)));
-    addStructAmpersandVisualizerAction                  = new QAction(QString("\"&&%1\"").arg(item->text(0)));
+    addVariableLoggerExpressionAction                   = new QAction(QString("%1").arg(item->text(0)));
+    addVariableLoggerAsteriskExpressionAction           = new QAction(QString("*%1").arg(item->text(0)));
+    addVariableLoggerAmpersandExpressionAction          = new QAction(QString("&&%1").arg(item->text(0)));
+    addVariableLoggerAsteriskAmpersandExpressionAction  = new QAction(QString("*&&%1").arg(item->text(0)));
+    addVariableLoggerObjcExpressionAction               = new QAction(QString("(objc)%1").arg(item->text(0)));
+    addVariableTrackerExpressionAction                  = new QAction(QString("%1").arg(item->text(0)));
+    addVariableTrackerAsteriskExpressionAction          = new QAction(QString("*%1").arg(item->text(0)));
+    addVariableTrackerAmpersandExpressionAction         = new QAction(QString("&&%1").arg(item->text(0)));
+    addVariableTrackerAsteriskAmpersandExpressionAction = new QAction(QString("*&&%1").arg(item->text(0)));
+    addVariableTrackerObjcExpressionAction              = new QAction(QString("(objc)%1").arg(item->text(0)));
+    addMemoryVisualizerAction                           = new QAction(QString("%1").arg(item->text(0)));
+    addMemoryAsteriskVisualizerAction                   = new QAction(QString("*%1").arg(item->text(0)));
+    addMemoryAmpersandVisualizerAction                  = new QAction(QString("&&%1").arg(item->text(0)));
+    addArrayVisualizerAction                            = new QAction(QString("%1").arg(item->text(0)));
+    addArrayAsteriskVisualizerAction                    = new QAction(QString("*%1").arg(item->text(0)));
+    addArrayAmpersandVisualizerAction                   = new QAction(QString("&&%1").arg(item->text(0)));
+    addMatrixVisualizerAction                           = new QAction(QString("%1").arg(item->text(0)));
+    addMatrixAsteriskVisualizerAction                   = new QAction(QString("*%1").arg(item->text(0)));
+    addMatrixAmpersandVisualizerAction                  = new QAction(QString("&&%1").arg(item->text(0)));
+    addStructVisualizerAction                           = new QAction(QString("%1").arg(item->text(0)));
+    addStructAsteriskVisualizerAction                   = new QAction(QString("*%1").arg(item->text(0)));
+    addStructAmpersandVisualizerAction                  = new QAction(QString("&&%1").arg(item->text(0)));
 
     QMenu menu("Visualizers", this);
     menu.setTitle("Visualizers");
@@ -197,6 +205,7 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
     loggerMenu.addAction(addVariableLoggerAsteriskExpressionAction);
     loggerMenu.addAction(addVariableLoggerAmpersandExpressionAction);
     loggerMenu.addAction(addVariableLoggerAsteriskAmpersandExpressionAction);
+    loggerMenu.addAction(addVariableLoggerObjcExpressionAction);
     menu.addMenu(&loggerMenu);
 
     QMenu trackerMenu("Add variable to Tracker");
@@ -204,6 +213,7 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
     trackerMenu.addAction(addVariableTrackerAsteriskExpressionAction);
     trackerMenu.addAction(addVariableTrackerAmpersandExpressionAction);
     trackerMenu.addAction(addVariableTrackerAsteriskAmpersandExpressionAction);
+    trackerMenu.addAction(addVariableTrackerObjcExpressionAction);
     menu.addMenu(&trackerMenu);
 
     QMenu memoryVisualizerMenu("Add variable to a Memory Visualizer");
@@ -290,6 +300,17 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
         return;
     }
 
+    // Handle adding a variable to log.
+    if (action == addVariableLoggerObjcExpressionAction) {
+
+        // Emit the signals.
+        if (variable != "") {
+            emit addVariableLoggerExpression(QString("(objc)") + variable);
+        }
+
+        return;
+    }
+
     // Handle adding a variable to track.
     if (action == addVariableTrackerExpressionAction) {
 
@@ -332,6 +353,18 @@ void SeerStackLocalsBrowserWidget::handleContextMenu (const QPoint& pos) {
         // Emit the signals.
         if (variable != "") {
             emit addVariableTrackerExpression(QString("*&") + variable);
+            emit refreshVariableTrackerValues();
+        }
+
+        return;
+    }
+
+    // Handle adding a variable to track.
+    if (action == addVariableTrackerObjcExpressionAction) {
+
+        // Emit the signals.
+        if (variable != "") {
+            emit addVariableTrackerExpression(QString("(objc)") + variable);
             emit refreshVariableTrackerValues();
         }